Moved "messages" package into "actors" one
