+++ client/src/main/scala/it/cwmp/client/controller/actors/ApiClientActor.scala
@@ -1,10 +1,10 @@
 package it.cwmp.client.controller.actors
 
 import akka.actor.Actor
-import it.cwmp.client.controller.messages.AuthenticationRequests.{LogIn, SignUp}
-import it.cwmp.client.controller.messages.AuthenticationResponses.{LogInFailure, LogInSuccess, SignUpFailure, SignUpSuccess}
-import it.cwmp.client.controller.messages.RoomsRequests._
-import it.cwmp.client.controller.messages.RoomsResponses._
+import it.cwmp.client.controller.actors.messages.AuthenticationRequests.{LogIn, SignUp}
+import it.cwmp.client.controller.actors.messages.AuthenticationResponses.{LogInFailure, LogInSuccess, SignUpFailure, SignUpSuccess}
+import it.cwmp.client.controller.actors.messages.RoomsRequests._
+import it.cwmp.client.controller.actors.messages.RoomsResponses._
 import it.cwmp.services.wrapper.{AuthenticationApiWrapper, RoomsApiWrapper}
 import it.cwmp.utils.Utils.stringToOption
