+++ client/src/main/scala/it/cwmp/client/controller/actors/messages/AuthenticationRequests.scala
@@ -1,4 +1,4 @@
-package it.cwmp.client.controller.messages
+package it.cwmp.client.controller.actors.messages
 
 /**
   * Collection of Authentication request messages
