+++ client/src/main/scala/it/cwmp/client/controller/actors/messages/Common.scala
@@ -1,4 +1,4 @@
-package it.cwmp.client.controller.messages
+package it.cwmp.client.controller.actors.messages
 
 /**
   * Identifies a request message
