+++ client/src/main/scala/it/cwmp/client/controller/actors/messages/RoomsResponses.scala
@@ -1,4 +1,4 @@
-package it.cwmp.client.controller.messages
+package it.cwmp.client.controller.actors.messages
 
 /**
   * The collection of Rooms response messages
