+++ client/src/test/scala/it/cwmp/client/controller/actors/ApiClientActorTest.scala
@@ -2,7 +2,7 @@ package it.cwmp.client.controller.actors
 
 import akka.actor.{ActorRef, ActorSystem, Props}
 import akka.testkit.{ImplicitSender, TestKit}
-import it.cwmp.client.controller.messages.{AuthenticationRequests, AuthenticationResponses, RoomsRequests, RoomsResponses}
+import it.cwmp.client.controller.actors.messages.{AuthenticationRequests, AuthenticationResponses, RoomsRequests, RoomsResponses}
 import it.cwmp.model.Address
 import it.cwmp.services.wrapper.{FakeAuthenticationApiWrapper, FakeRoomsApiWrapper}
 import it.cwmp.utils.Utils
